cleanup connectivities_as_numpy
