+++ model/testing/src/icon4py/model/testing/fixtures/__init__.py
@@ -9,4 +9,3 @@
 from __future__ import annotations
 
 from .datatest import *  # noqa: F403
-from .general import *  # noqa: F403
